Implemented manual overrides and more state logic
plus accelerometer

+++ include/kalmanFilter.h
@@ -4,19 +4,18 @@
 #include <stdint.h>
 #include <stdbool.h>
 #include <stdlib.h>
-#include "mpu9250.h"
 
 typedef struct {
-  double q; //process noise covariance
-  double r; //measurement noise covariance
-  double x; //value
-  double p; //estimation error covariance
-  double k; //kalman gain
+  int q; //process noise covariance
+  int r; //measurement noise covariance
+  int x; //value
+  int p; //estimation error covariance
+  int k; //kalman gain
 } kalman_state;
 
 
-kalman_state kalman_init(double q, double r, double p, double intial_value);
-double kalman_update(kalman_state* state, double measurement);
+kalman_state kalman_init(int q, int r, int p, int intial_value);
+int kalman_update(kalman_state* state, int measurement);
 void kalmanFunc(void);
